+++ packages/assets-controllers/src/MultichainAssetsRatesController/MultichainAssetsRatesController.ts
@@ -4,16 +4,17 @@ import type {
   AccountsControllerGetSelectedMultichainAccountAction,
 } from '@metamask/accounts-controller';
 import type {
-  RestrictedMessenger,
   ControllerStateChangeEvent,
   ControllerGetStateAction,
-} from '@metamask/base-controller';
+  StateMetadata,
+} from '@metamask/base-controller/next';
 import { type CaipAssetType, isEvmAccountType } from '@metamask/keyring-api';
 import type {
   KeyringControllerLockEvent,
   KeyringControllerUnlockEvent,
 } from '@metamask/keyring-controller';
 import type { InternalAccount } from '@metamask/keyring-internal-api';
+import type { Messenger } from '@metamask/messenger';
 import { StaticIntervalPollingController } from '@metamask/polling-controller';
 import type { HandleSnapRequest } from '@metamask/snaps-controllers';
 import type {
@@ -143,12 +144,10 @@ export type AllowedEvents =
 /**
  * Messenger type for the MultichainAssetsRatesController.
  */
-export type MultichainAssetsRatesControllerMessenger = RestrictedMessenger<
+export type MultichainAssetsRatesControllerMessenger = Messenger<
   typeof controllerName,
   MultichainAssetsRatesControllerActions | AllowedActions,
-  MultichainAssetsRatesControllerEvents | AllowedEvents,
-  AllowedActions['type'],
-  AllowedEvents['type']
+  MultichainAssetsRatesControllerEvents | AllowedEvents
 >;
 
 /**
@@ -158,17 +157,17 @@ export type MultichainAssetsRatesPollingInput = {
   accountId: string;
 };
 
-const metadata = {
+const metadata: StateMetadata<MultichainAssetsRatesControllerState> = {
   conversionRates: {
     includeInStateLogs: false,
     persist: true,
-    anonymous: true,
+    includeInDebugSnapshot: true,
     usedInUi: true,
   },
   historicalPrices: {
     includeInStateLogs: false,
     persist: false,
-    anonymous: true,
+    includeInDebugSnapshot: true,
     usedInUi: true,
   },
 };
@@ -213,7 +212,7 @@ export class MultichainAssetsRatesController extends StaticIntervalPollingContro
    * @param options - Constructor options.
    * @param options.interval - The polling interval in milliseconds.
    * @param options.state - The initial state.
-   * @param options.messenger - A reference to the messaging system.
+   * @param options.messenger - A reference to the messenger.
    */
   constructor({
     interval = 18000,
@@ -237,22 +236,22 @@ export class MultichainAssetsRatesController extends StaticIntervalPollingContro
     this.setIntervalLength(interval);
 
     // Subscribe to keyring lock/unlock events.
-    this.messagingSystem.subscribe('KeyringController:lock', () => {
+    this.messenger.subscribe('KeyringController:lock', () => {
       this.#isUnlocked = false;
     });
-    this.messagingSystem.subscribe('KeyringController:unlock', () => {
+    this.messenger.subscribe('KeyringController:unlock', () => {
       this.#isUnlocked = true;
     });
 
-    ({ accountsAssets: this.#accountsAssets } = this.messagingSystem.call(
+    ({ accountsAssets: this.#accountsAssets } = this.messenger.call(
       'MultichainAssetsController:getState',
     ));
 
-    ({ currentCurrency: this.#currentCurrency } = this.messagingSystem.call(
+    ({ currentCurrency: this.#currentCurrency } = this.messenger.call(
       'CurrencyRateController:getState',
     ));
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'CurrencyRateController:stateChange',
       async (currentCurrency: string) => {
         this.#currentCurrency = currentCurrency;
@@ -262,7 +261,7 @@ export class MultichainAssetsRatesController extends StaticIntervalPollingContro
         currencyRateControllerState.currentCurrency,
     );
 
-    this.messagingSystem.subscribe(
+    this.messenger.subscribe(
       'MultichainAssetsController:accountAssetListUpdated',
       async ({ assets }) => {
         const newAccountAssets = Object.entries(assets).map(
@@ -313,9 +312,7 @@ export class MultichainAssetsRatesController extends StaticIntervalPollingContro
    * @returns An array of internal accounts.
    */
   #listMultichainAccounts(): InternalAccount[] {
-    return this.messagingSystem.call(
-      'AccountsController:listMultichainAccounts',
-    );
+    return this.messenger.call('AccountsController:listMultichainAccounts');
   }
 
   /**
@@ -577,11 +574,9 @@ export class MultichainAssetsRatesController extends StaticIntervalPollingContro
 
       const selectedAccount =
         account ??
-        this.messagingSystem.call(
-          'AccountsController:getSelectedMultichainAccount',
-        );
+        this.messenger.call('AccountsController:getSelectedMultichainAccount');
       try {
-        const historicalPricesResponse = await this.messagingSystem.call(
+        const historicalPricesResponse = await this.messenger.call(
           'SnapController:handleRequest',
           {
             snapId: selectedAccount?.metadata.snap?.id as SnapId,
@@ -736,7 +731,7 @@ export class MultichainAssetsRatesController extends StaticIntervalPollingContro
   async #handleSnapRequest(args: SnapRequestArgs<unknown>): Promise<unknown> {
     const { snapId, handler, params } = args;
     try {
-      return await this.messagingSystem.call('SnapController:handleRequest', {
+      return await this.messenger.call('SnapController:handleRequest', {
         snapId,
         origin: 'metamask',
         handler,