+++ packages/assets-controllers/src/TokenSearchDiscoveryDataController/TokenSearchDiscoveryDataController.test.ts
@@ -1,5 +1,12 @@
-import { Messenger, deriveStateFromMetadata } from '@metamask/base-controller';
+import { deriveStateFromMetadata } from '@metamask/base-controller/next';
 import { ChainId } from '@metamask/controller-utils';
+import {
+  MOCK_ANY_NAMESPACE,
+  Messenger,
+  type MessengerActions,
+  type MessengerEvents,
+  type MockAnyNamespace,
+} from '@metamask/messenger';
 import type { Hex } from '@metamask/utils';
 import assert from 'assert';
 import { useFakeTimers } from 'sinon';
@@ -9,8 +16,6 @@ import {
   TokenSearchDiscoveryDataController,
   controllerName,
   MAX_TOKEN_DISPLAY_DATA_LENGTH,
-  type AllowedActions,
-  type AllowedEvents,
   type TokenSearchDiscoveryDataControllerMessenger,
   type TokenSearchDiscoveryDataControllerState,
 } from './TokenSearchDiscoveryDataController';
@@ -32,7 +37,11 @@ jest.mock('../token-service', () => {
   };
 });
 
-type MainMessenger = Messenger<AllowedActions, AllowedEvents>;
+type AllActions = MessengerActions<TokenSearchDiscoveryDataControllerMessenger>;
+
+type AllEvents = MessengerEvents<TokenSearchDiscoveryDataControllerMessenger>;
+
+type RootMessenger = Messenger<MockAnyNamespace, AllActions, AllEvents>;
 
 /**
  * Builds a not found token display data object.
@@ -111,13 +120,21 @@ function buildFoundTokenDisplayData(
  * @returns The restricted messenger.
  */
 function buildTokenSearchDiscoveryDataControllerMessenger(
-  messenger: MainMessenger = new Messenger(),
+  messenger: RootMessenger = new Messenger({ namespace: MOCK_ANY_NAMESPACE }),
 ): TokenSearchDiscoveryDataControllerMessenger {
-  return messenger.getRestricted({
-    name: controllerName,
-    allowedActions: ['CurrencyRateController:getState'],
-    allowedEvents: [],
+  const tokenSearchDiscoveryDataControllerMessenger = new Messenger<
+    typeof controllerName,
+    AllActions,
+    AllEvents,
+    RootMessenger
+  >({
+    namespace: controllerName,
   });
+  messenger.delegate({
+    messenger: tokenSearchDiscoveryDataControllerMessenger,
+    actions: ['CurrencyRateController:getState'],
+  });
+  return tokenSearchDiscoveryDataControllerMessenger;
 }
 
 /**
@@ -206,7 +223,9 @@ async function withController<ReturnValue>(
     callback = maybeCallback;
   }
 
-  const messenger = new Messenger<AllowedActions, AllowedEvents>();
+  const messenger: RootMessenger = new Messenger({
+    namespace: MOCK_ANY_NAMESPACE,
+  });
 
   messenger.registerActionHandler('CurrencyRateController:getState', () => ({
     currentCurrency: 'USD',
@@ -899,7 +918,7 @@ describe('TokenSearchDiscoveryDataController', () => {
           deriveStateFromMetadata(
             controller.state,
             controller.metadata,
-            'anonymous',
+            'includeInDebugSnapshot',
           ),
         ).toMatchInlineSnapshot(`Object {}`);
       });